GCP_pipeline_project/fastq_job/requirements.txt
google-cloud-storage==2.17.0